+++ _integrations/nrnats/nrnats_test.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"sync"
 	"testing"
+	"time"
 
 	"github.com/nats-io/nats-server/test"
 	"github.com/nats-io/nats.go"
@@ -24,6 +25,13 @@ func testApp(t *testing.T) newrelic.Application {
 	cfg.TransactionTracer.SegmentThreshold = 0
 	cfg.TransactionTracer.Threshold.IsApdexFailing = false
 	cfg.TransactionTracer.Threshold.Duration = 0
+	cfg.Attributes.Include = append(cfg.Attributes.Include,
+		newrelic.AttributeMessageRoutingKey,
+		newrelic.AttributeMessageQueueName,
+		newrelic.AttributeMessageExchangeType,
+		newrelic.AttributeMessageReplyTo,
+		newrelic.AttributeMessageCorrelationID,
+	)
 	app, err := newrelic.NewApplication(cfg)
 	if nil != err {
 		t.Fatal(err)
@@ -79,10 +87,8 @@ func TestStartPublishSegmentBasic(t *testing.T) {
 	app.(internal.Expect).ExpectMetrics(t, []internal.WantMetric{
 		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/all", Scope: "", Forced: false, Data: nil},
 		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/allOther", Scope: "", Forced: false, Data: nil},
-		{Name: "External/all", Scope: "", Forced: true, Data: nil},
-		{Name: "External/allOther", Scope: "", Forced: true, Data: nil},
-		{Name: "External/127.0.0.1:4222/all", Scope: "", Forced: false, Data: nil},
-		{Name: "External/127.0.0.1:4222/NATS/Publish/mysubject", Scope: "OtherTransaction/Go/testing", Forced: false, Data: nil},
+		{Name: "MessageBroker/NATS/Topic/Produce/Named/mysubject", Scope: "", Forced: false, Data: nil},
+		{Name: "MessageBroker/NATS/Topic/Produce/Named/mysubject", Scope: "OtherTransaction/Go/testing", Forced: false, Data: nil},
 		{Name: "OtherTransaction/Go/testing", Scope: "", Forced: true, Data: nil},
 		{Name: "OtherTransaction/all", Scope: "", Forced: true, Data: nil},
 		{Name: "OtherTransactionTotalTime", Scope: "", Forced: true, Data: nil},
@@ -100,11 +106,9 @@ func TestStartPublishSegmentBasic(t *testing.T) {
 		},
 		{
 			Intrinsics: map[string]interface{}{
-				"category":  "http",
-				"component": "NATS",
-				"name":      "External/127.0.0.1:4222/NATS/Publish/mysubject",
-				"parentId":  internal.MatchAnything,
-				"span.kind": "client",
+				"category": "generic",
+				"name":     "MessageBroker/NATS/Topic/Produce/Named/mysubject",
+				"parentId": internal.MatchAnything,
 			},
 			UserAttributes:  map[string]interface{}{},
 			AgentAttributes: map[string]interface{}{},
@@ -120,7 +124,7 @@ func TestStartPublishSegmentBasic(t *testing.T) {
 				Attributes:  map[string]interface{}{"exclusive_duration_millis": internal.MatchAnything},
 				Children: []internal.WantTraceSegment{
 					{
-						SegmentName: "External/127.0.0.1:4222/NATS/Publish/mysubject",
+						SegmentName: "MessageBroker/NATS/Topic/Produce/Named/mysubject",
 						Attributes:  map[string]interface{}{},
 					},
 				},
@@ -151,9 +155,9 @@ func TestSubWrapper(t *testing.T) {
 	}
 	wg := sync.WaitGroup{}
 	app := testApp(t)
-	nc.Subscribe("subject2", WgWrapper(&wg, SubWrapper(app, func(msg *nats.Msg) {})))
+	nc.QueueSubscribe("subject2", "queue1", WgWrapper(&wg, SubWrapper(app, func(msg *nats.Msg) {})))
 	wg.Add(1)
-	nc.Publish("subject2", []byte("data"))
+	nc.Request("subject2", []byte("data"), time.Second)
 	wg.Wait()
 
 	app.(internal.Expect).ExpectMetrics(t, []internal.WantMetric{
@@ -161,23 +165,38 @@ func TestSubWrapper(t *testing.T) {
 		{Name: "OtherTransactionTotalTime", Scope: "", Forced: true, Data: nil},
 		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/all", Scope: "", Forced: false, Data: nil},
 		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/allOther", Scope: "", Forced: false, Data: nil},
-		{Name: "OtherTransaction/Go/Message/NATS/Topic/subject2:subscriber", Scope: "", Forced: true, Data: nil},
-		{Name: "OtherTransactionTotalTime/Go/Message/NATS/Topic/subject2:subscriber", Scope: "", Forced: false, Data: nil},
+		{Name: "OtherTransaction/Go/Message/NATS/Topic/Named/subject2", Scope: "", Forced: true, Data: nil},
+		{Name: "OtherTransactionTotalTime/Go/Message/NATS/Topic/Named/subject2", Scope: "", Forced: false, Data: nil},
+	})
+	app.(internal.Expect).ExpectTxnEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":     "OtherTransaction/Go/Message/NATS/Topic/Named/subject2",
+				"guid":     internal.MatchAnything,
+				"priority": internal.MatchAnything,
+				"sampled":  internal.MatchAnything,
+				"traceId":  internal.MatchAnything,
+			},
+			AgentAttributes: map[string]interface{}{
+				"message.replyTo":    internal.MatchAnything, // starts with _INBOX
+				"message.routingKey": "subject2",
+				"message.queueName":  "queue1",
+			},
+			UserAttributes: map[string]interface{}{},
+		},
 	})
 }
 
-func TestStartPublishSegmentProcedure(t *testing.T) {
+func TestStartPublishSegmentNaming(t *testing.T) {
 	testCases := []struct {
-		subject   string
-		procedure string
+		subject string
+		metric  string
 	}{
-		{subject: "", procedure: "Publish"},
-		{subject: "mysubject", procedure: "Publish/mysubject"},
-		{subject: "_INBOX.asldfkjsldfjskd.ldskfjls", procedure: "Publish/_INBOX"},
+		{subject: "", metric: "MessageBroker/NATS/Topic/Produce/Named/Unknown"},
+		{subject: "mysubject", metric: "MessageBroker/NATS/Topic/Produce/Named/mysubject"},
+		{subject: "_INBOX.asldfkjsldfjskd.ldskfjls", metric: "MessageBroker/NATS/Topic/Produce/Temp"},
 	}
 
-	app := testApp(t)
-	txn := app.StartTransaction("testing", nil, nil)
 	nc, err := nats.Connect(nats.DefaultURL)
 	if nil != err {
 		t.Fatal(err)
@@ -185,10 +204,21 @@ func TestStartPublishSegmentProcedure(t *testing.T) {
 	defer nc.Close()
 
 	for _, tc := range testCases {
-		seg := StartPublishSegment(txn, nc, tc.subject)
-		if seg.Procedure != tc.procedure {
-			t.Errorf("incorrect Procedure:\nactual=%s\nexpected=%s", seg.Procedure, tc.procedure)
-		}
+		app := testApp(t)
+		txn := app.StartTransaction("testing", nil, nil)
+		StartPublishSegment(txn, nc, tc.subject).End()
+		txn.End()
+
+		app.(internal.Expect).ExpectMetrics(t, []internal.WantMetric{
+			{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/all", Scope: "", Forced: false, Data: nil},
+			{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/allOther", Scope: "", Forced: false, Data: nil},
+			{Name: "OtherTransaction/Go/testing", Scope: "", Forced: true, Data: nil},
+			{Name: "OtherTransaction/all", Scope: "", Forced: true, Data: nil},
+			{Name: "OtherTransactionTotalTime", Scope: "", Forced: true, Data: nil},
+			{Name: "OtherTransactionTotalTime/Go/testing", Scope: "", Forced: false, Data: nil},
+			{Name: tc.metric, Scope: "", Forced: false, Data: nil},
+			{Name: tc.metric, Scope: "OtherTransaction/Go/testing", Forced: false, Data: nil},
+		})
 	}
 }